Cleanup R initialization code.
Windows-specific initialization move to common module.
